more type checks
