+++ content/src/controller/routes.ts
@@ -16,12 +16,12 @@ import {
   getEntityThumbnail,
   getERC721Entity,
   getFailedDeployments,
-  getPointerChangesHandler,
   getStatus,
   getEntities
 } from './Controller'
 import { errorHandler } from './error-handler'
 import { filterByUrnHandler } from './filter-by-urn-handler'
+import { getPointerChangesHandler } from './pointer-changes-handler'
 
 // We return the entire router because it will be easier to test than a whole server
 export async function setupRouter({ components }: GlobalContext): Promise<Router<GlobalContext>> {
